# Bulan 1-2/Minggu 5-6/Project/inventory-stock-management-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inventory Stock Management API",
        "description": "REST API untuk mengelola data inventaris, produk, supplier, warehouse, dan stok.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/api/stock-in": {
            "post": {
                "tags": [
                    "Stock"
                ],
                "summary": "Stock In",
                "operationId": "531a3d78e6207ce47d91ba98b00374fe",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id",
                                    "quantity"
                                ],
                                "properties": {
                                    "product_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "note": {
                                        "type": "string",
                                        "example": "Barang masuk"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Stock In success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/stock-out": {
            "post": {
                "tags": [
                    "Stock"
                ],
                "summary": "Stock Out",
                "operationId": "1fd2dae0e5d911259355e4791e8497cd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id",
                                    "quantity"
                                ],
                                "properties": {
                                    "product_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "note": {
                                        "type": "string",
                                        "example": "Barang keluar"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Stock Out success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/stock-adjustment": {
            "post": {
                "tags": [
                    "Stock"
                ],
                "summary": "Stock Adjustment",
                "operationId": "43d64687699e68040b120735173706fb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id",
                                    "stock"
                                ],
                                "properties": {
                                    "product_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "note": {
                                        "type": "string",
                                        "example": "Stock opname"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Stock adjusted successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/stock-history": {
            "get": {
                "tags": [
                    "Stock"
                ],
                "summary": "Stock History",
                "operationId": "9ec256033470c0f6f450e3c395646036",
                "responses": {
                    "200": {
                        "description": "List Stock History"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/suppliers": {
            "get": {
                "tags": [
                    "Supplier"
                ],
                "summary": "Get All Suppliers",
                "operationId": "6207e57dd1dcd1d6e7125fe144607196",
                "responses": {
                    "200": {
                        "description": "List of suppliers"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Supplier"
                ],
                "summary": "Create Supplier",
                "operationId": "dc5dc7ccb140cd5b81cab6f42b396691",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "PT Dahana"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "081234567890"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Subang, Jawa Barat"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Supplier created successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/suppliers/{id}": {
            "get": {
                "tags": [
                    "Supplier"
                ],
                "summary": "Get Supplier Detail",
                "operationId": "7fc447855461518ce36bdeb5a44310a3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Supplier detail"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Supplier"
                ],
                "summary": "Update Supplier",
                "operationId": "22145931fdda85497cf90c73c5406747",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "PT Dahana Indonesia"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "081234567890"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Bandung"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Supplier updated successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Supplier"
                ],
                "summary": "Delete Supplier",
                "operationId": "2fe04361f0f1581a9f5dbb4094b634b2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Supplier deleted successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}